ranking float

+++ db/migrate/20201213131853_create_rankings.rb
@@ -1,7 +1,7 @@
 class CreateRankings < ActiveRecord::Migration[6.0]
   def change
     create_table :rankings do |t|
-      t.integer :point, null:false, default: '0'
+      t.float :point, null:false, default: '0'
       t.references :user, null: false, foreign_key: true
  
       t.timestamps
